Hide the class TopoStruct
