+++ oneflow/core/graph/straighten_nodes.cpp
@@ -22,6 +22,47 @@ limitations under the License.
 
 namespace oneflow {
 
+enum TaskClassifier : int {
+  kWaitingTransfer = 0,
+  kWaitingComputation = 1,
+  kRunASAP = 2,
+  kRunALAP = 3
+};
+
+class TopoStruct {
+ public:
+  TaskNode* node;
+  int32_t min_layer = -1;
+  int32_t tributary_layer = -1;
+  bool on_mainstem = false;
+  int32_t counter = 0;
+  int32_t min_distance2transfer = -1;
+  TopoStruct* next_same_node = nullptr;
+  // We can have some other nodes in it for example
+  // SbpNode<NdSbpSignature>* node;
+  // SbpEdge<NdSbpSignature>* node;
+  // Or we can omit all the pointers and leave all the useful parameters.
+
+  // Drop down the tributary layer
+  void DropTributaryLayer(int32_t upper_bound);
+
+  void SpreadTributaryLayer(HashMap<TaskNode*, TopoStruct>* task_node2topo_struct);
+
+  void SpreadMainstem(HashMap<TaskNode*, TopoStruct>* task_node2topo_struct);
+
+  // The minimum computation distance from the beginning of this op to the next transfer
+  int32_t GetMinDistance2Transfer(HashMap<TaskNode*, TopoStruct>* task_node2topo_struct);
+
+  // deciding parameter
+  // i = 0: those with small tributary layers go first
+  // i = 1: those with small minimum distance to transfer go first
+  // i = 2: first in first out
+  // i = 3: those with large tributary layers go first
+  // i = 4: those with long distance to transfer go first
+  // i = 5: last in first out
+  int32_t GetDecidingParameter(int32_t i) const;
+};
+
 namespace {
 
 // move the head from source to target
